Handle messages

+++ src/marine-licences/api/helpers/mas/worker-processor.js
@@ -1,16 +1,34 @@
 import { config } from '../../../../config.js'
 import { parseMessageBody } from '../../../../shared/common/helpers/sqs/parse-message-body.js'
 import { deleteMasMessage } from './sqs-client.js'
+import { updateTransferredMarineLicence } from './update-licence.js'
 
 const discardMalformedMessage = 'Discarding malformed MAS message'
 
 export const processMasMessage = async (server, message) => {
-  const { logger } = server
+  const { db, logger } = server
   const { sqsQueueName } = config.get('mas')
 
   const body = parseMessageBody(message, logger, discardMalformedMessage)
-  if (body) {
-    logger.info({ body }, 'Received MAS message')
+
+  if (!body) {
+    await deleteMasMessage(sqsQueueName, message.ReceiptHandle)
+    return
+  }
+
+  logger.info({ body }, 'Received MAS message')
+
+  const { applicationReference, status } = body
+
+  if (!applicationReference) {
+    throw new Error('No Application Reference exists on message')
+  }
+
+  if (status === 'transferred') {
+    await updateTransferredMarineLicence(db, logger, {
+      body,
+      id: message.MessageId
+    })
   }
 
   await deleteMasMessage(sqsQueueName, message.ReceiptHandle)
